work in progress refactor converter add maori
